Basic "plugins" (visitor) feature
When usign the api, it is sometime necessary to act on the ast object
between csso operations. Instead of having to deal with options and calls to
parse and compress, this change makes it easy to edit the ast object
after csso's parse (afterParse) and compress (afterCompress) operation.

This is especially good for task runner, which can accept those
parameters to be able to speed task execution.